requirements.txt
# Core dependencies
qq-botpy>=1.2.0
openai>=1.0.0

aiohttp>=3.8.0
pyyaml>=6.0
psutil>=5.9.0

# Optional: Scheduler support
apscheduler>=3.10.0

# Optional: Vertex AI support
# google-cloud-aiplatform>=1.38.0

# Memory system
chromadb>=0.4.0
